pkg/pyproject.toml
[project]
name = "langchain_tableau"
version = "0.4.35"
description = "Tableau tools for Agentic use cases with Langchain"
authors = [
    { name = "Stephen Price" },
    { name = "Joe Constantino" },
    { name = "Joseph Fluckiger" }
]
requires-python = ">=3.12.2"
license = {text = "MIT"}
readme = "README.md"

dependencies = [
    "python-dotenv",
    "langchain",
    "langchain-core",
    "langgraph",
    "langchain-openai",
    "requests",
    "pydantic",
    "pyjwt",
    "aiohttp",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

[project.urls]
"Homepage" = "https://github.com/Tab-SE/tableau_langchain"
"Bug Tracker" = "https://github.com/Tab-SE/tableau_langchain/issues"

[tool.hatch.build.targets.wheel]
packages = ["langchain_tableau"]

[tool.hatch.build.targets.sdist]
include = [
    "langchain_tableau",
    "README.md",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
